requirements.txt
Flask
Inject
requests
paho-mqtt
